fix: 매물 페이지 병목현상 찾아서 성능 개선
