+++ src/app/store/[id]/page.tsx
@@ -28,7 +28,7 @@ const StoreDetailContent = dynamic(
       <Box
         sx={{
           width: '100%',
-          height: '100vh',
+          height: '50vh',
           display: 'flex',
           justifyContent: 'center',
           alignItems: 'center',
