+++ src/components/store/StoreWindow.tsx
@@ -1,5 +1,6 @@
 'use client';
 
+import { useState, useEffect } from 'react';
 import Box from '@mui/material/Box';
 import Button from '@mui/material/Button';
 import Card from '@mui/material/Card';
@@ -9,28 +10,36 @@ import Divider from '@mui/material/Divider';
 import Skeleton from '@mui/material/Skeleton';
 import Tooltip from '@mui/material/Tooltip';
 import Typography from '@mui/material/Typography';
+import { useQuery } from '@tanstack/react-query';
 import { useRouter } from 'next/navigation';
-import { StoreMinimum } from '@/api/store/getStoreMinimum';
+import { getStoreMinimum } from '@/api/store/getStoreMinimum';
+import { LOCALSTORAGE_RECENT_STORE_KEY } from '@/app/store/[id]/constants';
 import { DEFAULT_WINDOW_WIDTH } from '@/components/store/constants';
 import { convertMoneyString } from '@/utils';
 import { BoldLabelValue } from '../common/store-card/elements';
 
 interface StoreWindowProps {
-  storeData?: StoreMinimum[];
-  isLoading?: boolean;
   openModal?: () => void;
   handleStoreChange?: (newStoreName: string) => void;
   onCardClick?: () => void;
 }
 const StoreWindow = ({
-  storeData,
-  isLoading,
   openModal,
   handleStoreChange,
   onCardClick,
 }: StoreWindowProps) => {
   const router = useRouter();
-  const isEmptyData = !storeData || storeData.length === 0;
+  const [recentStores, setRecentStores] = useState<string[] | undefined>(
+    undefined,
+  );
+
+  const { data: storeData, isLoading } = useQuery({
+    queryKey: ['store-minimum', recentStores],
+    queryFn: () => getStoreMinimum({ idArr: recentStores }),
+    enabled: recentStores !== undefined,
+  });
+
+  const isEmptyData = !storeData || storeData.data?.length === 0;
 
   const handleCardClick = (storeId: string) => {
     onCardClick && onCardClick();
@@ -41,6 +50,16 @@ const StoreWindow = ({
     handleStoreChange && handleStoreChange(''); // window에서 열릴 때는 storename 초기화
     openModal && openModal();
   };
+
+  useEffect(() => {
+    const localStorageRecentStores = localStorage.getItem(
+      LOCALSTORAGE_RECENT_STORE_KEY,
+    );
+    const parsedRecentStores: string[] = localStorageRecentStores
+      ? JSON.parse(localStorageRecentStores)
+      : [];
+    setRecentStores(parsedRecentStores);
+  }, []);
 
   return (
     <Card
@@ -73,7 +92,8 @@ const StoreWindow = ({
         {isLoading && <Skeleton variant='text' height='50px' />}
         {!isLoading &&
           !isEmptyData &&
-          storeData.map((storeData, index) => (
+          storeData.data &&
+          storeData.data.map((storeData, index) => (
             <Tooltip
               title={storeData.store_name ?? ''}
               key={'recent-store' + storeData.store_id + index}
